Extract test run reconstruction from diagnostic case
